rename mfpackages.py to gwf.py,  see
#71
