+++ nlmod/mfpackages/__init__.py
@@ -1,2 +1,2 @@
 from . import constant_head, recharge, surface_water, horizontal_flow_barrier
-from .mfpackages import *
+from .gwf import *
